Unvendor lazy_loader (#5681)
Now that `lazy_loader` is being released as a package, it may be easier
to depend on it than to vendor the code.

@tlambert03 has implemented a feature in `lazy_loader` that allows lazy
loading, while also letting type checkers do their jobs. See
https://scientific-python.org/specs/spec-0001/#type-checkers

I could transition the napari lazy imports to the stub loader in a next
PR, if that's desired.